## Version 1.0.4.1 - 2025-11-12 (#15)
- WAInputNumber:
  * Step property was not passing through to the resulting input box
  * Changed step property to `decimal` to allow greater precision definition

+++ Components/WAInputNumber.razor.cs
@@ -117,7 +117,7 @@ namespace WebAwesomeBlazor.Components
         /// Default value is null.
         /// </remarks>
         [Parameter]
-        public double? Step { get; set; }
+        public decimal? Step { get; set; }
 
         /// <summary>
         /// The input's maximum value
